+++ parser_library/src/config/proc_grps.h
@@ -140,12 +140,23 @@ struct preprocessor_options
     friend bool operator==(const preprocessor_options&, const preprocessor_options&) = default;
 };
 
+struct external_function
+{
+    std::string name;
+    std::variant<int32_t, std::string> value;
+
+    friend bool operator==(const external_function&, const external_function&) = default;
+};
+void to_json(nlohmann::json& j, const std::vector<external_function>& p);
+void from_json(const nlohmann::json& j, std::vector<external_function>& p);
+
 struct processor_group
 {
     std::string name;
     std::vector<std::variant<library, dataset, endevor, endevor_dataset>> libs;
     assembler_options asm_options;
     std::vector<preprocessor_options> preprocessors;
+    std::optional<std::vector<external_function>> external_functions;
 };
 void to_json(nlohmann::json& j, const processor_group& p);
 void from_json(const nlohmann::json& j, processor_group& p);
@@ -154,6 +165,7 @@ struct proc_grps
 {
     std::vector<processor_group> pgroups;
     std::vector<std::string> macro_extensions;
+    std::optional<std::vector<external_function>> external_functions;
 };
 void to_json(nlohmann::json& j, const proc_grps& p);
 void from_json(const nlohmann::json& j, proc_grps& p);